pyproject.toml
[project]
name = "operantid"
version = "0.1.1"
description = "Universal autonomous AI agent library for high-performance browser navigation and automated reasoning."
authors = [
    {name = "Junior_7i", email = "antojunimaia@gmail.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["ai", "agent", "browser", "automation", "llm", "playwright", "gemini", "openai", "mistral", "autonomous"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.6",
    "mistralai>=1.0.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
    "flask>=3.0.0"
]

[project.urls]
Homepage = "https://github.com/antojunior/operantid"
Repository = "https://github.com/antojunior/operantid"
"Bug Tracker" = "https://github.com/antojunior/operantid/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
